+++ MdeModulePkg/Application/VgaShim/Int10hHandler.h
