+++ MdeModulePkg/Application/VgaShim/Int10hHandler.sh
